Now when Composite ID (a Primary Key made from more than one column) that also references (have a foreign key) another table, we add all columns in the same "References()" call.
References(x => x.AcaartCodGen).Columns("acaart_CodGen, acaart_CodEle1, acaart_CodEle2, acaart_CodEle3");

Bing for "reference composite fluentnhibernate"
or read
http://devlicio.us/blogs/derik_whittaker/archive/2009/01/30/mapping-a-one-to-many-relationship-with-fluent-nhibernate-w-composite-id-s.aspx

as it says: "The use of WithColumns() on the References (many-to-one) mapping when trying to map back to the parent.  This will tell NHibernate to use the composite-id as the mapping key."

This is buggy yet, but so far it works.